requirements.txt
tornado==5.1.1
tormysql==0.4.0
requests==2.20.1
PyMySQL==0.9.2
PyJWT==1.6.4
PyYAML==3.13
Pillow==5.3.0
pytesseract==0.2.5
redis==3.0.1